// src/app/personajesw/interfaces/sw.interfaces.ts
// Generated by https://quicktype.io

export interface Swapi {
  count:    number;
  next:     string;
  previous: null;
  results:  Swresult[];
}

export interface Swresult {
  name:       string;
  height:     string;
  mass:       string;
  hair_color: string;
  skin_color: string;
  eye_color:  string;
  birth_year: string;
  gender:     Gender;
  homeworld:  string;
  films:      string[];
  species:    string[];
  vehicles:   string[];
  starships:  string[];
  created:    string;
  edited:     string;
  url:        string;
}

export enum Gender {
  Female = "female",
  Male = "male",
  NA = "n/a",
}

export interface People{
  id: string,
  name:string,
  mass: string,
  birth_year:String,
  gender:Gender,
  eye_color:String

}